refactor: split summary_normalizer.go into modular files (fix #36 TL review)
Split the 843-line summary_normalizer.go into 4 focused files:
- normalizer.go (123 lines) — Core NormalizeToCanonical and NormalizeFromStructuredArtifacts entry points
- normalizer_sections.go (173 lines) — Section extraction (H1, metadata block, H2 sections)
- normalizer_tables.go (273 lines) — Table parsing (action items, risks)
- normalizer_lists.go (272 lines) — List parsing (decisions, open issues, discussion sections)

All files now under 500 lines threshold. All 48+ tests still pass.

References: #36
